+++ Libft/srcs/str/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: svet <svet@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/18 13:14:50 by skrasin           #+#    #+#             */
-/*   Updated: 2020/07/17 18:28:26 by svet             ###   ########.fr       */
+/*   Updated: 2020/09/01 18:57:07 by svet             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,5 +15,5 @@
 
 int	ft_atoi(const char *str)
 {
-	return (ft_strtol(str, NULL, 10));
+	return ((int)ft_strtol(str, NULL, 10));
 }
