+++ Libft/srcs/str/ft_strmapi.c
@@ -6,18 +6,18 @@
 /*   By: svet <svet@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/20 19:37:24 by skrasin           #+#    #+#             */
-/*   Updated: 2020/07/21 18:35:03 by svet             ###   ########.fr       */
+/*   Updated: 2020/09/01 19:52:19 by svet             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_string.h"
 #include <sys/_types/_null.h>
 
-char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
+char	*ft_strmapi(char const *s, char (*f)(size_t, char))
 {
-	char			*fstr;
-	char			*fstrcpy;
-	unsigned int	n;
+	char	*fstr;
+	char	*fstrcpy;
+	size_t	n;
 
 	n = 0;
 	if (s == NULL || (fstr = ft_strnew(ft_strlen(s))) == NULL)
