+++ Libft/srcs/str/ft_strtoul.c
@@ -6,7 +6,7 @@
 /*   By: svet <svet@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/14 18:50:13 by svet              #+#    #+#             */
-/*   Updated: 2020/08/14 18:54:37 by svet             ###   ########.fr       */
+/*   Updated: 2020/09/01 19:50:48 by svet             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,5 +14,5 @@
 
 unsigned long	ft_strtoul(const char *nptr, char **endptr, int base)
 {
-	return (ft_strtol(nptr, endptr, base));
+	return ((unsigned long)ft_strtol(nptr, endptr, base));
 }
